% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely_detect_changepoints.R
\name{cost}
\alias{cost}
\title{Evaluating a cost of specified data segment}
\usage{
cost(data, u, v, sumstat, K)
}
\arguments{
\item{data}{vector of numeric values}

\item{u}{index of the segment beginning}

\item{v}{index of the segment end}

\item{sumstat}{matrix of statistics used to calculate empirical distribution function}

\item{K}{the number of quantiles used in integral approximation, not greater than
the number of all observations in the dataset}
}
\value{
a cost of specified data segment
}
\description{
The cost() function evaluates cost for given data and both beginning and
end of the segment.
}
