% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely_detect_interactions.R
\name{interaction_measure}
\alias{interaction_measure}
\title{Detecting interactions via a permutation approach}
\usage{
interaction_measure(explainer, var1, var2, inter_param)
}
\arguments{
\item{explainer}{DALEX explainer created with explain() function}

\item{var1}{character, first of the names of variables for which wa want to evaluate
interaction effect}

\item{var2}{character, second of the names of variables for which wa want to evaluate
interaction effect}

\item{inter_param}{numeric, a positive value indicating which of single observation non-additive effects
are to be regarded as significant, the higher value the higher non-additive effect has to be to be taken
into account}
}
\value{
vector of logical values, each of them corresponding to one observation and
indicating whether for this observation interaction effect was detected
}
\description{
The safely_detect_interactions() function detects second-order interactions based on predictions
made by a surrogate model. For each pair of features it performs values permutation in order
to evaluate their non_additive effect.
}
