% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely_detect_changepoints.R
\name{PELT_algorithm}
\alias{PELT_algorithm}
\title{PELT algorithm}
\usage{
PELT_algorithm(data, penalty_value, sumstat, K)
}
\arguments{
\item{data}{vector of numeric values}

\item{penalty_value}{penalty value returned by penalty_value_choice() function}

\item{sumstat}{matrix of statistics used to calculate empirical distribution function}

\item{K}{the number of quantiles used in integral approximation, not greater than
the number of all observations in the dataset}
}
\description{
The PELT_algorithm function() performs PELT algorithm in order to find
optimal changepoints.
}
