% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely_detect_changepoints.R
\name{penalty_value_choice}
\alias{penalty_value_choice}
\title{Choosing penalty value}
\usage{
penalty_value_choice(penalty, n)
}
\arguments{
\item{penalty}{penalty for introducing another changepoint,
one of "AIC", "BIC", "SIC", "MBIC", "Hannan-Quinn" or numeric non-negative value}

\item{n}{integer, number of observations}
}
\value{
penalty value
}
\description{
The penalty_value_choice() function sets penalty value.
}
