% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_extraction.R
\name{plot.safe_extractor}
\alias{plot.safe_extractor}
\title{Plotting transformations of the SAFE extractor object}
\usage{
\method{plot}{safe_extractor}(x, ..., variable = NULL)
}
\arguments{
\item{x}{safe_extractor object containing information about variables transformations created with safe_extraction() function}

\item{...}{other parameters}

\item{variable}{character, name of the variable to be plotted}
}
\value{
a plot object
}
\description{
Plotting transformations of the SAFE extractor object
}
