% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely_transform_continuous.R
\name{pretty_intervals}
\alias{pretty_intervals}
\title{Creating good-looking intervals from breakpoints}
\usage{
pretty_intervals(break_points)
}
\arguments{
\item{break_points}{a vector of breakpoints from which you wish to create intervals}
}
\value{
list of character intervals
}
\description{
The pretty_intervals() function creates intervals from given breakpoints.
If possible, it formats numbers using prettyNum() function to make them look better when printing.
}
