% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_extraction.R
\name{print.safe_extractor}
\alias{print.safe_extractor}
\title{Printing Summary of the SAFE Extractor Object}
\usage{
\method{print}{safe_extractor}(x, ..., variable = NULL)
}
\arguments{
\item{x}{safe_extractor object containing information about variables transformations created with safe_extraction() function}

\item{...}{other parameters}

\item{variable}{character, name of the variable to be plotted. If this argument
is not specified then transformations for all variables are printed}
}
\value{
No return value, prints the structure of the object
}
\description{
Printing Summary of the SAFE Extractor Object
}
