% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFE-package.R
\docType{package}
\name{SAFE-package}
\alias{SAFE}
\alias{SAFE-package}
\title{SAFE: Surrogate-assisted feature extraction}
\description{
Surrogate-assisted feature extraction
}
\author{
\strong{Maintainer}: Anna Gierlak \email{anna.gierlak@op.pl}

}
\keyword{internal}
