% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely_select_variables.R
\name{safely_select_variables}
\alias{safely_select_variables}
\title{Selecting variables}
\usage{
safely_select_variables(safe_extractor, data, y = NULL, which_y = NULL,
  encoding = "categorical", verbose = TRUE)
}
\arguments{
\item{safe_extractor}{object containing information about variables transformations created with safe_extractor() function}

\item{data}{data, with already transformed variables or not, containing response variable or not}

\item{y}{vector of responses, must be given if data does not contain it}

\item{which_y}{numeric or character (optional), must be given if data contains response values}

\item{encoding}{method of representing factor variables, one of: "categorical", "one-hot"}

\item{verbose}{logical, if progress bar is to be printed}
}
\value{
vector of variables names, selected based on AIC values
}
\description{
The safely_select_variables() function selects variables from dataset returned
by transform_data(). For each original variable:
\enumerate{
\item if encoding = "categorical" then exactly one variable is chosen
}
\itemize{
\item either original one or transformed one. The choice is based on the AIC value
for linear model (regression) or logistic regression (classification).
}
\enumerate{
\item if encoding = "one-hot" then backward feature selection is performed
on all features
}
}
