% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely_transform_factor.R
\name{safely_transform_factor}
\alias{safely_transform_factor}
\title{Factor feature transformation using factorMerger package}
\usage{
safely_transform_factor(explainer, variable)
}
\arguments{
\item{explainer}{DALEX explainer created with explain() function}

\item{variable}{a feature for which the transformation function is to be computed}
}
\value{
list of information on the transformation of given variable
}
\description{
The safely_transform_factor() function calculates a transformation function
for the factor variable using the getOptimalPartitionDf() function from factorMerger package.
}
