% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safer.R
\name{safer}
\alias{safer}
\title{Creating SAFE-extractor}
\usage{
safer(explainer, package = "changepoint.np", response_type = "pdp")
}
\arguments{
\item{explainer}{DALEX explainer created with explain() function}

\item{package}{character, a package used to compute breakpoints in case of continuous variables,
one of: "changepoint.np", "strucchange"}

\item{response_type}{character, type of response to be calculated, one of: "pdp", "ale".
If features are uncorrelated, one can use "pdp" type - otherwise "ale" is strongly recommended.}
}
\value{
SAFE-extractor object containing information about variables transformations
}
\description{
The safer() function creates a SAFE-extractor object which may be used later
for surrogate feature extraction.
}
