% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_continuous.R
\name{transform_continuous}
\alias{transform_continuous}
\title{Continuous feature transformation using PDP plot}
\usage{
transform_continuous(explainer, variable, package = "changepoint.np",
  type = "constant", plot = FALSE)
}
\arguments{
\item{explainer}{DALEX explainer created with explain() function}

\item{variable}{a feature for which the transformation function is to be computed}

\item{package}{a package used to compute breakpoints, one of: "changepoint.np", "strucchange"}

\item{type}{a transformation method, one of: "constant", "linear"}

\item{plot}{logical, if TRUE then plots breakpoints and segments means}
}
\value{
transformation function for the given variable
}
\description{
The transform_continuous() function calculates a transformation function
for the continuous variable using a PDP plot from DALEX package.
}
