% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Features transformation using PDP plots}
\usage{
transform_data(data, propositions, user_interaction = FALSE,
  keep_old = FALSE)
}
\arguments{
\item{data}{data for which features values are to be transformed}

\item{propositions}{list of transformations functions and interactions returned by transform_propositons() function}

\item{user_interaction}{logical, TRUE allows the user to decide whether to keep the old variables after transformation}

\item{keep_old}{logical, if FALSE then old variables are removed from the dataset}
}
\value{
data with extra columns containing transformed variables
}
\description{
The transform_data() function creates new variables
using PDP plots from DALEX package and adds them to the original ones.
}
