% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_factor.R
\name{transform_factor}
\alias{transform_factor}
\title{Factor feature transformation using factorMerger package}
\usage{
transform_factor(explainer, variable, plot = FALSE)
}
\arguments{
\item{explainer}{DALEX explainer created with explain() function}

\item{variable}{a feature for which the transformation function is to be computed}

\item{plot}{logical, if TRUE then plots the optimal partition}
}
\value{
transformation function for the given variable
}
\description{
The transform_factor() function calculates a transformation function
for the factor variable using the getOptimalPartitionDf() function from factorMerger package.
}
