% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_propositions.R
\name{transform_propositions}
\alias{transform_propositions}
\title{Features transformation using PDP plots}
\usage{
transform_propositions(explainer, package = "changepoint.np",
  type = "constant", plot = FALSE)
}
\arguments{
\item{explainer}{DALEX explainer created with explain() function}

\item{package}{a package used to compute breakpoints, one of: "changepoint.np", "strucchange"}

\item{type}{a transformation method, one of: "constant", "linear"}

\item{plot}{logical, if TRUE then plots breakpoints and segments means}
}
\value{
list of propositions of variables transformations
}
\description{
The transform_propositions() function computes transformation functions
for all continuous variables using PDP plots from DALEX package.
}
