% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_variables.R
\name{transform_variables}
\alias{transform_variables}
\title{Features transformation using PDP plots}
\usage{
transform_variables(explainer, data, package = "changepoint.np",
  type = "constant", plot = FALSE)
}
\arguments{
\item{explainer}{DALEX explainer created with explain() function}

\item{data}{data for which features values are to be transformed}

\item{package}{a package used to compute breakpoints, one of: "changepoint.np", "strucchange"}

\item{type}{a transformation method, one of: "constant", "linear"}

\item{plot}{logical, if TRUE then plots breakpoints and segments means}
}
\value{
data with extra columns containing transformed variables
}
\description{
The transform_variables() function creates new variables
using PDP plots from DALEX package and adds them to the original ones.
}
