% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSAFE-package.R
\docType{package}
\name{rSAFE-package}
\alias{rSAFE}
\alias{rSAFE-package}
\title{rSAFE: Surrogate-Assisted Feature Extraction}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a model agnostic tool for white-box model trained on features extracted from a black-box model. For more information see: Gosiewska et al. (2020) <arXiv:2002.04267>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ModelOriented/rSAFE}
  \item Report bugs at \url{https://github.com/ModelOriented/rSAFE/issues}
}

}
\author{
\strong{Maintainer}: Alicja Gosiewska \email{alicjagosiewska@gmail.com}

Authors:
\itemize{
  \item Anna Gierlak \email{anna.gierlak@op.pl}
  \item Przemyslaw Biecek [thesis advisor]
}

Other contributors:
\itemize{
  \item Michal Burdukiewicz (\href{https://orcid.org/0000-0001-8926-582X}{ORCID}) [contributor]
}

}
\keyword{internal}
