% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSAFE-package.R
\docType{package}
\name{rSAFE-package}
\alias{rSAFE}
\alias{rSAFE-package}
\title{rSAFE: Surrogate-Assisted Feature Extraction in R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides a model agnostic tool for making an explainable
  white-box model more accurate using alternative black-box model called surrogate model.
  Based on the complicated model, such as neural network or random forest, new features
  are being extracted and then used in the process of fitting a simpler explainable model,
  improving its overall performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ModelOriented/rSAFE}
  \item Report bugs at \url{https://github.com/ModelOriented/rSAFE/issues}
}

}
\author{
\strong{Maintainer}: Alicja Gosiewska \email{alicjagosiewska@gmail.com}

Authors:
\itemize{
  \item Anna Gierlak \email{anna.gierlak@op.pl}
  \item Przemyslaw Biecek [thesis advisor]
}

}
\keyword{internal}
