% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFE-package.R
\docType{package}
\name{SAFE-package}
\alias{SAFE}
\alias{SAFE-package}
\title{SAFE: Surrogate-assisted feature extraction}
\description{
Surrogate-assisted feature extraction
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MI2DataLab/SAFE}
  \item Report bugs at \url{https://github.com/MI2DataLab/SAFE/issues}
}

}
\author{
\strong{Maintainer}: Anna Gierlak \email{anna.gierlak@op.pl}

Authors:
\itemize{
  \item Przemyslaw Biecek [thesis advisor]
}

}
\keyword{internal}
