% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFE-package.R
\docType{package}
\name{SAFE-package}
\alias{SAFE}
\alias{SAFE-package}
\title{SAFE: Surrogate-assisted feature extraction}
\description{
The SAFE package is a model agnostic tool for making an interpretable
  white-box model more accurate using alternative black-box model called surrogate model.
  Based on the complicated model, such as neural network or random forest, new features
  are being extracted and then used in the process of fitting a simpler interpretable model,
  improving its overall performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MI2DataLab/SAFE}
  \item Report bugs at \url{https://github.com/MI2DataLab/SAFE/issues}
}

}
\author{
\strong{Maintainer}: Anna Gierlak \email{anna.gierlak@op.pl}

Authors:
\itemize{
  \item Przemyslaw Biecek [thesis advisor]
}

}
\keyword{internal}
