% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely_transform_continuous.R
\name{safely_transform_continuous}
\alias{safely_transform_continuous}
\title{Continuous feature transformation using PDP plot}
\usage{
safely_transform_continuous(variable, explainer, response_type = "ale",
  penalty = "MBIC", no_segments = 2)
}
\arguments{
\item{variable}{a feature for which the transformation function is to be computed}

\item{explainer}{DALEX explainer created with explain() function}

\item{response_type}{character, type of response to be calculated, one of: "pdp", "ale".
If features are uncorrelated, one can use "pdp" type - otherwise "ale" is strongly recommended.}

\item{penalty}{penalty for introducing another changepoint,
one of "AIC", "BIC", "SIC", "MBIC", "Hannan-Quinn" or numeric non-negative value}

\item{no_segments}{numeric, a number of segments variable is to be divided into in case of founding no breakpoints}
}
\value{
list of information on the transformation of given variable
}
\description{
The safely_transform_continuous() function calculates a transformation function
for the continuous variable using a PDP plot obtained from black box model.
}
