% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely_transform_data.R
\name{safely_transform_data}
\alias{safely_transform_data}
\title{Features transformation using PDP plots}
\usage{
safely_transform_data(safe_extractor, data, encoding = "categorical",
  verbose = TRUE)
}
\arguments{
\item{safe_extractor}{object containing information about variables transformations created with safe_extractor() function}

\item{data}{data for which features are to be transformed}

\item{encoding}{method of representing factor variables, one of: "categorical", "one-hot"}

\item{verbose}{logical, if progress bar is to be printed}
}
\value{
data with extra columns containing newly created variables
}
\description{
The transform_data() function creates new variables in dataset
using SAFE-extractor object.
}
