% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_continuous.R
\name{transform_continuous}
\alias{transform_continuous}
\title{Continuous feature transformation using PDP plot}
\usage{
transform_continuous(variable, explainer, package = "changepoint.np",
  response_type = "pdp")
}
\arguments{
\item{variable}{a feature for which the transformation function is to be computed}

\item{explainer}{DALEX explainer created with explain() function}

\item{package}{character, a package used to compute breakpoints, one of: "changepoint.np", "strucchange"}

\item{response_type}{character, type of response to be calculated, one of: "pdp", "ale".
If features are uncorrelated, one can use "pdp" type - otherwise "ale" is strongly recommended.}
}
\value{
list of transformation information on the given variable
}
\description{
The transform_continuous() function calculates a transformation function
for the continuous variable using a PDP plot from DALEX package.
}
