% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Features transformation using PDP plots}
\usage{
transform_data(data, safe_extractor, encoding = "categorical")
}
\arguments{
\item{data}{data for which features are to be transformed}

\item{safe_extractor}{object containing information about variables transformations created with safer() function}

\item{encoding}{method of representing factor variables, one of: "categorical", "one-hot"}
}
\value{
data with extra columns containing newly created variables
}
\description{
The transform_data() function creates new variables in dataset
using SAFE-extractor object.
}
