% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Features transformation using PDP plots}
\usage{
transform_data(data, y, propositions, which_variables = "only_new")
}
\arguments{
\item{data}{data for which features values are to be transformed}

\item{propositions}{list of transformations functions and interactions returned by transform_propositons() function}

\item{which_variables}{allows user to decide which variables keep, one of: "all", "only_new", "aic"}
}
\value{
data with extra columns containing transformed variables
}
\description{
The transform_data() function creates new variables
using PDP plots from DALEX package and adds them to the original ones.
}
